requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Pillow == 9.2.0
bs4 == 0.0.1
config == 0.5.1
fuzzywuzzy == 0.18.0
matplotlib == 3.5.2
nltk == 3.7
numpy == 1.23.0
pandas == 1.4.3
plotly == 5.9.0
requests == 2.28.1
scikit_learn == 1.1.1
scipy == 1.8.1
sentence_transformers == 2.2.2
spacy == 3.3.1
streamlit == 1.10.0
streamlit_player == 0.1.5
streamlit_timeline == 0.0.2
transformers == 4.20.1
wordcloud == 1.8.2.2
yake == 0.4.8
